+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Implemented `.to_bits()` and `.from_bits()` for `APyCFloat`.
-- LaTeX representation for `APyCFloat`.
+- LaTeX representation for `APyCFixed` and `APyCFloat`.
 
 ### Fixed